Ajustes e build dos assets
