+++ package.json
@@ -6,8 +6,17 @@
         "build": "vite build"
     },
     "devDependencies": {
+        "autoprefixer": "^10.4.19",
         "axios": "^1.6.4",
         "laravel-vite-plugin": "^1.0.0",
+        "postcss": "^8.4.38",
+        "scrollreveal": "^4.0.9",
+        "tailwindcss": "^3.4.3",
         "vite": "^5.0.0"
+    },
+    "dependencies": {
+        "@tailwindcss/aspect-ratio": "^0.4.2",
+        "@tailwindcss/forms": "^0.5.7",
+        "@tailwindcss/typography": "^0.5.12"
     }
 }
