--- public/css/style-qr-code.css
@@ -1,12 +1,0 @@
-/* Estilo para a div que contém o QR code */
-.qr-code-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-/* Estilo para o SVG dentro da div width="897px" height="452px"  */
-.qr-code-container svg {
-    width: 897px;
-    height: 452px;
-}
