+++ resources/views/qrcodes/login.blade.php
@@ -7,17 +7,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>{{env('app_name')}}</title>
     <meta name="description" content="Sistema de gestão de freezers - Revolucionando a forma como você compra bebidas">
-
-    <script src="https://cdn.tailwindcss.com"></script>
-    <script src="https://cdn.tailwindcss.com?plugins=forms,typography,aspect-ratio,line-clamp"></script>
-    <link href="{{asset('css/style-qr-code.css')}}" rel="stylesheet">
+    @vite('resources/css/app.css')
+    @vite('resources/js/app.js')
 </head>
 
 <body class="leading-normal tracking-normal text-gray-900" style="font-family: 'Source Sans Pro', sans-serif;">
-<div class="h-screen pb-14 bg-right bg-cover" style="background-image:url({{asset('images/landing-page/bg.svg')}});">
+<div id="bg-cover">
     <!--Nav-->
     <div class="w-full container mx-auto p-6">
-
         <div class="w-full flex items-center justify-between">
             <a class="flex items-center text-indigo-400 no-underline hover:no-underline font-bold text-2xl lg:text-4xl logo"  href="{{route('frontend.home')}}">
                 <img src="{{asset('images/brands/icon-340.png')}}" class="mr-3 h-6 sm:h-9" alt="{{env('APP_NAME')}}"/> {{env('APP_NAME')}}
@@ -37,46 +34,31 @@
                     <svg class="fill-current h-7" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 20 20"><path fill-rule="evenodd" d="M10 .333A9.911 9.911 0 0 0 6.866 19.65c.5.092.678-.215.678-.477 0-.237-.01-1.017-.014-1.845-2.757.6-3.338-1.169-3.338-1.169a2.627 2.627 0 0 0-1.1-1.451c-.9-.615.07-.6.07-.6a2.084 2.084 0 0 1 1.518 1.021 2.11 2.11 0 0 0 2.884.823c.044-.503.268-.973.63-1.325-2.2-.25-4.516-1.1-4.516-4.9A3.832 3.832 0 0 1 4.7 7.068a3.56 3.56 0 0 1 .095-2.623s.832-.266 2.726 1.016a9.409 9.409 0 0 1 4.962 0c1.89-1.282 2.717-1.016 2.717-1.016.366.83.402 1.768.1 2.623a3.827 3.827 0 0 1 1.02 2.659c0 3.807-2.319 4.644-4.525 4.889a2.366 2.366 0 0 1 .673 1.834c0 1.326-.012 2.394-.012 2.72 0 .263.18.572.681.475A9.911 9.911 0 0 0 10 .333Z" clip-rule="evenodd"/></svg>
                 </a>
             </div>
-
         </div>
-
     </div>
-
     <!--Main-->
     <div class="container pt-24 md:pt-48 px-6 mx-auto flex flex-wrap flex-col md:flex-row items-center">
-
         <!--Left Col-->
         <div class="flex flex-col w-full xl:w-2/5 justify-center lg:items-start overflow-y-hidden description">
             <h1 class="my-4 text-3xl md:text-5xl text-purple-800 font-bold leading-tight text-center md:text-left slide-in-bottom-h1">Revolucionando a forma como você compra bebidas</h1>
             <p class="leading-normal text-base md:text-2xl mb-8 text-center md:text-left slide-in-bottom-subtitle">Não perca tempo, aproveite agora as nossas deliciosas bebidas!</p>
 
-            <a href="{{ route('filament.app.auth.login') }}" type="button" class="shadow-lg button text-2xl text-blue-600 text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800">
+            <a href="{{ route('filament.app.auth.login') }}" type="button" class="text-center shadow-lg button text-2xl text-blue-600 text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg px-5 py-2.5 me-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800">
                 Acesse agora!
             </a>
 
         </div>
-
         <!--Right Col-->
         <div class="w-full xl:w-3/5 py-0 overflow-y-hidden">
             <div class="qr-code-container qr-code">
                 {!! html_entity_decode($qrcode) !!}
             </div>
         </div>
-
-
         <!--Footer-->
-        <div class="w-full pt-16 pb-6 text-sm text-center md:text-left fade-in">
+        <div class="w-full pt-16 pb-6 text-sm text-center md:text-left fade-in footer">
             <a class="text-gray-500 no-underline hover:no-underline" href="#">{{env('app_name')}} &copy; App {{date('Y')}}</a>
         </div>
-
     </div>
-
-
 </div>
-    <!--=======================  SCROLLREVEAL  =======================-->
-    <script src="https://unpkg.com/scrollreveal"></script>
-    <!--=======================  JS  =======================-->
-    <script type="text/javascript" rel="script" src="{{asset('js/script-scrollreveal.js')}}"></script>
 </body>
-
 </html>
