+++ vite.config.js
@@ -1,5 +1,6 @@
 import { defineConfig } from 'vite';
 import laravel from 'laravel-vite-plugin';
+import tailwindcss from "tailwindcss";
 
 export default defineConfig({
     plugins: [
@@ -8,4 +9,13 @@ export default defineConfig({
             refresh: true,
         }),
     ],
+    css: {
+        postcss: {
+            plugins: [
+                tailwindcss({
+                    config: "./tailwind.config.js",
+                }),
+            ],
+        },
+    },
 });
